Refactors structure.
Refactors GUI to be contained in one file (tapestry.py) in the classes folder.
