--- sdk/src/scrapegoat/classes/tapestry/control_panel.py
@@ -1,6 +1,0 @@
-from textual.containers import HorizontalGroup
-from textual.widgets import Button, Label
-
-class ControlPanel(HorizontalGroup):
-	def compose(self):
-		yield Label("<no node selected>")
